moveForward header compatible across sim and robot code

+++ software/Mouse/Algo/API.c
@@ -51,8 +51,8 @@ int wallLeft() {
     return getBoolean("wallLeft");
 }
 
-int moveForward(int spaces) {
-    printf("moveForward %d\n", spaces);
+int moveForward(double spaces) {
+    printf("moveForward %d\n", (int)spaces);
     fflush(stdout);
     char response[BUFFER_SIZE];
     fgets(response, BUFFER_SIZE, stdin);
